refactor: introduce BaseExtendedAndReadonlyAttributeParser for shared parsing logic
Created BaseExtendedAndReadonlyAttributeParser to consolidate common parsing logic, including methods for resolving and validating large size value symbols and handling constructor access modifiers. Refactored ExtendedBitFieldAttributeParser to inherit from the new base class, improving code organization and reusability.
