+++ PropertyBitPack.SourceGen/AttributeParsers/ExtendedBitFieldAttributeParser.cs
@@ -19,10 +19,8 @@ namespace PropertyBitPack.SourceGen.AttributeParsers;
 /// <see cref="ExtendedBitFieldAttribute"/>. It ensures proper diagnostics are reported for missing 
 /// or invalid configurations, such as type mismatches or missing large size value references.
 /// </remarks>
-internal sealed class ExtendedBitFieldAttributeParser : BaseAttributeParser
+internal sealed class ExtendedBitFieldAttributeParser : BaseExtendedAndReadonlyAttributeParser
 {
-    private static readonly Type[] _validSymbolsTypes = [typeof(IFieldSymbol), typeof(IPropertySymbol), typeof(IMethodSymbol)];
-
     /// <summary>
     /// Determines if the specified attribute is a candidate for parsing.
     /// </summary>
@@ -79,157 +77,5 @@ internal sealed class ExtendedBitFieldAttributeParser : BaseAttributeParser
         return true;
     }
 
-    /// <summary>
-    /// Attempts to resolve the symbol for the large size value referenced in the attribute.
-    /// </summary>
-    /// <param name="semanticModel">The semantic model used for analysis.</param>
-    /// <param name="propertySyntax">The property being analyzed.</param>
-    /// <param name="attributeData">The attribute data being parsed.</param>
-    /// <param name="attributeSyntax">The syntax representation of the attribute.</param>
-    /// <param name="owner">The owner type of the property.</param>
-    /// <param name="diagnostics">
-    /// A collection of <see cref="Diagnostic"/> instances to which any parsing errors are added.
-    /// </param>
-    /// <param name="getterLargeSizeValueSymbol">The resolved symbol for the large size value.</param>
-    /// <returns>
-    /// <c>true</c> if the large size value symbol was successfully resolved; otherwise, <c>false</c>.
-    /// </returns>
-    private bool TryGetterLargeSizeValueSymbol(SemanticModel semanticModel, PropertyDeclarationSyntax propertySyntax, AttributeData attributeData, AttributeSyntax attributeSyntax, INamedTypeSymbol owner, ref readonly  ImmutableArrayBuilder<Diagnostic> diagnostics, [NotNullWhen(true)] out ISymbol? getterLargeSizeValueSymbol)
-    {
-        getterLargeSizeValueSymbol = null;
-
-        var argument = GetAttributeArgument(attributeSyntax, nameof(IExtendedBitFieldAttribute.GetterLargeSizeValueName));
-
-        // for diagnostics
-        var propertyName = propertySyntax.Identifier.Text;
-
-        if (argument is null)
-        {
-            var diagnostic = Diagnostic.Create(
-                PropertyBitPackDiagnostics.MissingGetterLargeSizeValue,
-                attributeSyntax.GetLocation(),
-                propertySyntax.Identifier.Text
-            );
-
-            diagnostics.Add(diagnostic);
-            return false;
-        }
-
-        var nameOfOperation = semanticModel.GetOperation(argument.Expression);
-
-        if (nameOfOperation is not INameOfOperation nameOf)
-        {
-            var diagnostic = Diagnostic.Create(
-                PropertyBitPackDiagnostics.MandatoryOfNameofInGetterLargeSizeValueName,
-                argument.GetLocation(),
-                propertyName
-            );
-
-            diagnostics.Add(diagnostic);
-
-            return false;
-        }
-
-        var validSymbolsTypes = _validSymbolsTypes.AsSpan();
-
-        getterLargeSizeValueSymbol = ExtractSymbolFromNameOf(owner, nameOf, validSymbolsTypes);
-
-        // Maybe we should use Unsafe.As<IPropertySymbol?>(getterLargeSizeValueSymbol) instead of casting?
-        var propertySymbol = (IPropertySymbol?)semanticModel.GetDeclaredSymbol(propertySyntax);
-
-
-        return ValidateAndReport(getterLargeSizeValueSymbol, propertySymbol, propertySyntax, argument, in diagnostics);
-    }
-
-    /// <summary>
-    /// Validates the resolved large size value symbol against the property type and reports diagnostics if invalid.
-    /// </summary>
-    /// <param name="getterLargeValueSymbol">The symbol for the large size value.</param>
-    /// <param name="propertySymbol">The property symbol being validated.</param>
-    /// <param name="propertyDeclarationSyntax">The syntax representation of the property.</param>
-    /// <param name="getterLargeValueSyntax">The syntax representation of the large size value reference.</param>
-    /// <param name="diagnostics">
-    /// A collection of <see cref="Diagnostic"/> instances to which any validation errors are added.
-    /// </param>
-    /// <returns>
-    /// <c>true</c> if validation succeeds; otherwise, <c>false</c>.
-    /// </returns>
-    private bool ValidateAndReport(ISymbol? getterLargeValueSymbol, IPropertySymbol? propertySymbol, PropertyDeclarationSyntax propertyDeclarationSyntax, AttributeArgumentSyntax getterLargeValueSyntax, ref readonly ImmutableArrayBuilder<Diagnostic> diagnostics)
-    {
-        if (getterLargeValueSymbol is null)
-        {
-            return false;
-        }
-
-        if (propertySymbol is null)
-        {
-            // This should never happen, but just in case
-            return false;
-        }
-
-        var getterLargeValueLocation = getterLargeValueSyntax.GetLocation();
-
-        if (getterLargeValueSymbol is IPropertySymbol property)
-        {
-            if (!SymbolEqualityComparer.Default.Equals(property.Type, propertySymbol.Type))
-            {
-                var diagnostic = Diagnostic.Create(
-                    PropertyBitPackDiagnostics.InvalidTypeForPropertyReference,
-                    getterLargeValueLocation,
-                    property.Name,
-                    propertySymbol.Type
-                );
-
-                diagnostics.Add(diagnostic);
-                
-                return false;
-            }
-        }
-        else if (getterLargeValueSymbol is IMethodSymbol method)
-        {
-            if (!SymbolEqualityComparer.Default.Equals(method.ReturnType, propertySymbol.Type))
-            {
-                var diagnostic = Diagnostic.Create(
-                    PropertyBitPackDiagnostics.InvalidReturnTypeForMethodReference,
-                    getterLargeValueLocation,
-                    method.Name,
-                    propertySymbol.Type,
-                    method.ReturnType
-                );
-                diagnostics.Add(diagnostic);
-                return false;
-            }
-        }
-        else if(getterLargeValueSymbol is IFieldSymbol fieldSymbol)
-        {
-            if(!SymbolEqualityComparer.Default.Equals(fieldSymbol.Type, propertySymbol.Type))
-            {
-                // I know this is a bit confusing, but sometimes we use a property to reference a field
-                // thats why we report it as a property not a field (because we doesn't have DiagnosticDescriptor for fields)
-
-                // TODO: Add a new DiagnosticDescriptor for fields
-                var diagnostic = Diagnostic.Create(
-                    PropertyBitPackDiagnostics.InvalidTypeForPropertyReference,
-                    getterLargeValueLocation,
-                    fieldSymbol.Name,
-                    propertySymbol.Type
-                );
-                diagnostics.Add(diagnostic);
-                return false;
-            }
-        }
-        else
-        {
-            var diagnostic = Diagnostic.Create(
-                PropertyBitPackDiagnostics.InvalidReferenceInGetterLargeSizeValueName,
-                propertySymbol.Locations[0],
-                propertySymbol.Name
-            );
-            diagnostics.Add(diagnostic);
-            return false;
-        }
-
-        return true;
-
-    }
+    
 }
